Web management interface
Create web management service for configuration.

Move all runtime state into var/ directory. setup_amity creates the directory along with the subdirectories log,
config, gui, gui/nicegui, and homekit.

Web management is a UI layer ontop of the existing CLI configuration tools. This means that config.yaml is
updated through the CLI tools. The only exception is the activity editor that saves activities directly to
the config - the equivalent of editing config.yaml with a text editor in the CLI.

All commands stop/edit config/start the hub (previously 'Amity') when updating config.yaml. In addition, the hub
monitors config.yaml for changes, and exits if any are detected. The assumption is that systemd will
restart it. Because the hub needs to be always on with the new UI, if config is insufficient for start, it
essentially hangs until config.yaml is updated. Before, it used to crash which would eventually lead to systemd
giving up on the service and not starting it anymore.