Wait for CEC adapters to initialize properly
If CEC adapter initialization fails it is almost certainly because the GPIO pin is not being pulled up
with a pullup resistor. If the onboard RPi pullups are set, then this will never happen. However, if
the GPIO pins are configured for Amity Board, and the board isn't connected, then there are no pullups, and
CEC adapter initialization will fail. Instead of failing out of the hub, and tickling systemd to eventually
give up on the service, it's better to sit and wait for the pin to be pulled up. So retry every 10 seconds.