Comprehensive, unified setup and image build scripts
Running build/build-image.sh (preferably on a beefy Ubuntu machine) compiles the required kernels,
and builds a ready image for writing to an SD card for immediate use on a RPi. Further Amity
configuration is done through the web interface.

For manual setup on an already running RPi, downloading, and running setup_amity may work much as before
but it is not tested, and will likely be dropped since maintaining the kernels separately online is extra
work that doesn't yield much benefit. Building a full image is very fast with the right hardware.

Replace setup_amity with configure_amity for enable/disable of Amity.