Samsung SolarCell (Smart) Remote support
Samsung remotes show up as non-standard Linux input devices. They pretend to be a keyboard and a mouse, but
are completely non-standard devices. To accommodate using evdev inputs for both standard keyboards and
Samsung remotes, this change includes:

1. Generalizing the old keyboard.py to evdev_input.py for tracking input devices and using outside handlers
to identify (probe) and handle the devices' events.
2. Keyboard specific handling in keyboard.py in the form of a Handler class that probes for and
handles events from keyboards.
3. Adding solarcell.py that defines a Handler class that probes for and handles events from SolarCell remotes.