Long press and short press key mapping
In standby mode, power button short press is mapped to the first activity (select button activity). This is more intuitive
for most people who expect the power button to turn on the systemi from standby. This feature along with the select button
being the biggest button on the remote means that the first activity should ideally be configured to the most used activity on the
system. For most systems that is likely the media player. Power button long press is now standby; to shutdown any errant devices.

Activity selection button long presses are now mapped to add 5 more activities for a total of 10.
For example, long pressing select, selects the 6th activity, up selects the 7th activity, etc.

Short and long presses are taken into account in macros as well. So if currently in the 2nd activity, pressing the power button
and short pressing the select button will select the 1st activity. While pressing the power button and long pressing the
select button will select the 6th activity.

In addition, in an activity, short pressing the power button goes to standby, while long pressing the power button will fix
the activity.

Long presses are considered presses longer than 0.5 seconds