Emulate play/pause state to workaround play/pause problem
Most remotes have a single play/pause button. However, HDMI-CEC doesn't seem to have a play/pause key
code equivalent. Instead, HDMI-CEC has discrete play and pause key codes.

Until now, the workaround has been to map play/pause to the select key which behaves more or less as
expected in most cases. Particularly, when watching video. However, it fails when playing music on a media
player while browsing other elements of the UI that are not the playback control. Select will select
whatever is displaying in the UI instead of actually pausing or resuming the background music. This requires
navigating back to the playback control which can be onerous.

This change attempts to work around the issue by alternating play/pause key presses between the discrete
play and pause keys to the playback device. Since we don't know the actual play/pause state of the device,
the selected key can be wrong and do nothing. The assumption is that the user will press the key again
to alternate to the correct key code and that it won't be perceived as irritating flakiness.