Amity Fin v1.0
Refactor Amity Board into a fin that rises up and out from the Raspberry Pi header directly. The fin removes
the need for a wiring harness and makes it easier to put together the hardware.
The unusual form factor is an easy way to ensure that the HDMI connectors and cables are far enough away
from the Raspberry Pi board so that there is no interference.